backend/requirements.txt
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
PyYAML==6.0.1
loguru==0.7.2
rich==13.7.1
langchain==0.2.6
langchain-google-genai==1.0.7
langchain-groq==0.1.6
langchain-community==0.2.6
langchain-core==0.2.10
google-generativeai==0.7.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
tiktoken==0.7.0
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.0
numpy==1.26.4
pandas==2.1.4
joblib==1.4.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-multipart==0.0.9
sqlalchemy==2.0.30
pymysql==1.1.1
pymongo==4.7.2
mlflow==2.12.2
ragas==0.1.10
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
